[update] 펀딩 삭제/펀딩 참여 취소에 기간 제한 걸음

+++ src/main/java/com/habday/server/service/FundingService.java
@@ -222,7 +222,9 @@ public class FundingService extends Common {
         FundingItem fundingItem = fundingItemRepository.findById(fundingItemId)
                 .orElseThrow(() -> new CustomException(NO_FUNDING_ITEM_ID));
 
-
+        if(LocalDate.now().compareTo(fundingItem.getFinishDate())>=0){
+            throw new CustomException(DELETE_FUNDING_UNAVAILABLE);
+        }
         fundingItemRepository.delete(fundingItem);
         payService.unscheduleAll(fundingItem);
         emailFormats.sendFundingCanceledEmail(fundingItem);
